Altered approach to image embeddings for better performance
I slightly changed the collate function and late fusion model forward function to process each item's images as a separate batch instead of just concatenating all the images in each batch together to form one stacked tensor. Since batch['images'] now returns a list instead of a tensor, I then changed the creation of image embeddings in the forward function -- after each tensor in  the list is sent through the image encoder, I aggregate the embeddings by taking the mean. Then, I concatenate all the aggregated embeddings into a single tensor for that batch (to be concatenated with the audio and text embeddings). I found that by distinguishing the individual items of each batch in this way, the model performed better and I was able to do it without encountering shape/dimensionality issues.